queries/requirements.txt
streamlit>=1.31
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
pandas>=2.0